+++ docs/usage.rst
@@ -2,6 +2,6 @@
 Usage
 =====
 
-To use pytest-cover in a project::
+To use pytest-cov in a project::
 
-	import pytest_cover
+	import pytest_cov
