+++ setup.cfg
@@ -17,10 +17,10 @@ exclude = tests/*,*/migrations/*,*/south_migrations/*
 
 [bumpversion:file:docs/conf.py]
 
-[bumpversion:file:src/pytest_cover/__init__.py]
+[bumpversion:file:src/pytest_cov/__init__.py]
 
 [pytest]
-norecursedirs = 
+norecursedirs =
 	.git
 	.tox
 	.env
@@ -29,11 +29,11 @@ norecursedirs =
 	south_migrations
 	migrations
 	example
-python_files = 
+python_files =
 	test_*.py
 	*_test.py
 	tests.py
-addopts = 
+addopts =
 	-rxEfs
 	--strict
 	--ignore=docs/conf.py
@@ -47,21 +47,21 @@ addopts =
 [isort]
 force_single_line = True
 line_length = 120
-known_first_party = pytest_cover
+known_first_party = pytest_cov
 default_section = THIRDPARTY
 forced_separate = test_pytest_cover
 
 [matrix]
-python_versions = 
+python_versions =
 	2.6
 	2.7
 	3.3
 	3.4
 	pypy
-dependencies = 
-coverage_flags = 
+dependencies =
+coverage_flags =
 	: true
 	nocover: false
-environment_variables = 
+environment_variables =
 	-
