+++ tests/test_pytest_cover.py
@@ -7,7 +7,7 @@ import coverage
 import virtualenv
 import py
 import pytest
-import pytest_cover.plugin
+import pytest_cov.plugin
 
 pytest_plugins = 'pytester', 'cov'
 
@@ -584,7 +584,7 @@ def test_dist_boxed(testdir):
 
 
 def test_not_started_plugin_does_not_fail(testdir):
-    plugin = pytest_cover.plugin.CovPlugin(None, None, start=False)
+    plugin = pytest_cov.plugin.CovPlugin(None, None, start=False)
     plugin.pytest_sessionfinish(None, None)
     plugin.pytest_terminal_summary(None)
